modified godot card
made it so that the parent is a marker2d, with sprite2d and area2d children, so that the hitboxes and the sprite are actually seperate and the hitbox wont scale with the card.

we would also not need to do isflipping checks anymore

however, taking the mouse off during flipping still makes the x axis of the card as it flips back fucky. and it ends up looking stretched.

completely unnecessary at this point, but im gonna try to fix it.